New Alert view

+++ src/View/ListShapeView.java
@@ -6,11 +6,11 @@ import Domain.ShapeDB;
 import Domain.ShapeHolder;
 import javafx.collections.FXCollections;
 import javafx.geometry.Insets;
-import javafx.scene.control.Button;
-import javafx.scene.control.Label;
-import javafx.scene.control.ListView;
+import javafx.scene.control.*;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.VBox;
+
+import java.util.Optional;
 
 
 public class ListShapeView {
@@ -35,11 +35,13 @@ public class ListShapeView {
 
 //refresh list
         Button refreshButton = new Button( "Refresh List" );
+        refreshButton.setMinWidth( 120 );
         refreshButton.setOnAction( actionEvent -> {
             listViewNames.setItems( FXCollections.observableArrayList( shapeHolder.getAllNamesAndProperties() ) );
         } );
 
         Button deleteButton = new Button( "Delete Shape" );
+        deleteButton.setMinWidth( 120 );
         deleteButton.setOnAction( actionEvent -> {
             //get IndexNr from listView (deleteNr)
             int deleteNr = listViewNames.getSelectionModel().getSelectedIndex();
@@ -55,6 +57,7 @@ public class ListShapeView {
 
         //Add Shape to DB
         Button addShapeToDB = new Button ("Add shape to DB");
+        addShapeToDB.setMinWidth( 240 );
         addShapeToDB.setOnAction( actionEvent -> {
             //Convert shape into ShapeDB
             int shapeNr = listViewNames.getSelectionModel().getSelectedIndex();
@@ -65,18 +68,18 @@ public class ListShapeView {
             dao.save( shapeForDB );
         } );
 
-        //Delete Shape from DB
-        //Add Shape to DB
-        Button deleteShapeFromDB = new Button ("Delete shape from DB");
-        deleteShapeFromDB.setOnAction( actionEvent -> {
-            //Convert shape into ShapeDB
-            int shapeNr = listViewNames.getSelectionModel().getSelectedIndex();
-            //get Shape from indexNr
-            Shape shapeFromList = shapeHolder.getShape( shapeNr );
-            ShapeDB shapeForDB = new ShapeDB( shapeFromList.getNameOfShape(), shapeFromList.getRadius(), shapeFromList.getLength(), shapeFromList.getWidth(), shapeFromList.getHeight() );
-            DAO dao = new DAO();
-            dao.delete( shapeForDB );
-        } );
+//        //Delete Shape from DB
+//        //Add Shape to DB
+//        Button deleteShapeFromDB = new Button ("Delete shape from DB");
+//        deleteShapeFromDB.setOnAction( actionEvent -> {
+//            //Convert shape into ShapeDB
+//            int shapeNr = listViewNames.getSelectionModel().getSelectedIndex();
+//            //get Shape from indexNr
+//            Shape shapeFromList = shapeHolder.getShape( shapeNr );
+//            ShapeDB shapeForDB = new ShapeDB( shapeFromList.getNameOfShape(), shapeFromList.getRadius(), shapeFromList.getLength(), shapeFromList.getWidth(), shapeFromList.getHeight() );
+//            DAO dao = new DAO();
+//            dao.delete( shapeForDB );
+//        } );
 
         //Get objects from DB
         Button shapesFromDB = new Button( "Get all Shapes from DB" );
@@ -86,12 +89,18 @@ public class ListShapeView {
             //        listViewNames.setItems( FXCollections.observableArrayList( shapeHolder.getAllNamesAndProperties() ) );
         } );
 
-        //Get objects from DB
+        //Delete all from DB from DB
         Button deleteAllFromButton = new Button( "Delete all Shapes from DB" );
         deleteAllFromButton.setMinWidth( 240 );
         deleteAllFromButton.setOnAction( actionEvent -> {
-            DAO dao = new DAO();
-            dao.deleteAll();
+            Alert alert = new Alert( Alert.AlertType.WARNING, "Are you sure you want to empty the Database?", ButtonType.YES, ButtonType.NO);
+            Optional<ButtonType> result = alert.showAndWait();
+            if (result.isPresent() && result.get() == ButtonType.YES) {
+                DAO dao = new DAO();
+                dao.deleteAll();
+            } else {
+                alert.close();
+            }
         } );
 
         //DB Buttons in an Hbox
